Add nameval type to projectfile for shorthand asy usage
